move hint component
