+++ src/content/components/follow.js
@@ -1,6 +1,6 @@
 import * as followActions from 'content/actions/follow';
 import messages from 'shared/messages';
-import Hint from 'content/hint';
+import Hint from './hint';
 import HintKeyProducer from 'content/hint-key-producer';
 
 const DEFAULT_HINT_CHARSET = 'abcdefghijklmnopqrstuvwxyz';
