+++ src/content/components/hint.js
